# trader_interactive/public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "L5 Swagger API",
        "description": "L5 Swagger API description",
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "schemes": [
        "http"
    ],
    "paths": {
        "/api/listings": {
            "get": {
                "tags": [
                    "listings"
                ],
                "summary": "Get list of listings",
                "description": "Returns list of listings",
                "operationId": "getListings",
                "responses": {
                    "200": {
                        "description": "Listing Details Found"
                    },
                    "500": {
                        "description": "Listing Details Not Found"
                    }
                }
            }
        },
        "/api/listing-details/id={id}": {
            "get": {
                "tags": [
                    "listing-details"
                ],
                "summary": "Get Individual Listing Details",
                "description": "Returns Individual Listing Details",
                "operationId": "getListingDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "listing id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Individual Listing Details Found"
                    },
                    "500": {
                        "description": "Individual Listing Details Not Found"
                    }
                }
            }
        },
...
        "/api/sellers": {
            "get": {
                "tags": [
                    "sellers"
                ],
                "summary": "Get list of sellers",
                "description": "Returns list of sellers",
                "operationId": "getSellers",
                "responses": {
                    "200": {
                        "description": "Sellers Details Found"
                    },
                    "500": {
                        "description": "Sellers Details Not Found"
                    }
                }
            }
        },
        "/api/seller-details/id={id}": {
            "get": {
                "tags": [
                    "seller-details"
                ],
                "summary": "Get Individual Seller Details",
                "description": "Returns Individual Seller Details",
                "operationId": "getSellerDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "seller id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Individual Seller Details Found"
                    },
                    "500": {
                        "description": "Individual Seller Details Not Found"
                    }
                }
            }
        },
...
        "/api/seller-individual-listing-details/id={id}/type={type}": {
            "get": {
                "tags": [
                    "seller-individual-listing-details"
                ],
                "summary": "Get Seller Individual Listing Details",
                "description": "Returns Seller Individual Listing Details",
                "operationId": "getSellerListingDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "seller id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "type",
                        "in": "path",
                        "description": "listing type",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Seller Individual Listing Details Found"
                    },
                    "500": {
                        "description": "Seller Individual Listing Details Not Found"
                    }
                }
            }
        },
...
        "/api/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get list of users",
                "description": "Returns list of users",
                "operationId": "getUsers",
                "responses": {
                    "200": {
                        "description": "User Details Found"
                    },
                    "500": {
                        "description": "User Details Not Found"
                    }
                }
            }
        },
...
    },
...
}